Update AAD4_E1_Jeremy.java
El typecast se realiza al resultado de la división, NO al número antes de dividir, porque sino se cambia el tipo de número y el resultado perdería dígitos. Es importante tomar en cuenta.

+++ Codigo/AAD4_E1_Jeremy.java
@@ -70,7 +70,7 @@ public class AAD4_E1_Jeremy{
 		while (numerotemp >0)
 		{
 			// Resto = Obtener el Resto de la división NúmeroTemp  / 10
-			resto = numerotemp % 10;
+			resto = (int)(numerotemp % 10);
 				// Si el Resto = Número a contabilizar
 				if (resto == numeroacontabilizar)
 				{
